+++ CoreOS/system/kernel/purr_version.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#define PURR_OS_VERSION       "0.9.2"
+#define PURR_OS_VERSION       "0.9.4"
 #define PURR_OS_VERSION_MAJOR 0
 #define PURR_OS_VERSION_MINOR 9
-#define PURR_OS_VERSION_PATCH 2
+#define PURR_OS_VERSION_PATCH 4
 
 #define KITT_VERSION          "0.5.3"
 #define KITT_VERSION_MAJOR    0